'' black  used

+++ services/event_to_notification/src/models/bill.py
@@ -3,9 +3,10 @@ from models.base import FastJsonModel
 
 class IncomingBill(FastJsonModel):
     """Message to send model ."""
+
     bill_uuid: str
     status: str
     user_uuid: str
     type: str
     item_uuid: str
-    amount: int
+    amount: float
